Rename result-tables.tsx => ResultTables.tsx
