+++ extensions/ql-vscode/src/view/results/ResultTables.tsx
